Default thumbnail

+++ pyread/orm.py
@@ -1,4 +1,5 @@
 import os
+from flask import url_for
 from pyread import db, book_upload_set, cover_upload_set, utils
 
 
@@ -43,7 +44,10 @@ class Book(db.Model):
     genre = db.relationship('Genre')
 
     def get_thumb_url(self):
-        return cover_upload_set.url('thumb-' + self.cover)
+        if self.cover:
+            return cover_upload_set.url('thumb-' + self.cover)
+        else:
+            return url_for('static', filename='img/default.png')
 
     def get_format(self):
         if self.filename:
@@ -54,29 +58,31 @@ class Book(db.Model):
     def attempt_to_update_file(self, file):
         try:
             filename = book_upload_set.save(file)
-            if self.filename:
-                # remove current file if user uploaded new one
-                try:
-                    os.remove(book_upload_set.path(self.filename))
-                except:
-                    pass  # can't delete old book. not the end of the world.
+            if filename:
+                if self.filename:
+                    # remove current file if user uploaded new one
+                    try:
+                        os.remove(book_upload_set.path(self.filename))
+                    except:
+                        pass  # can't delete old book. not the end of the world.
 
-            self.filename = filename
+                self.filename = filename
         except:
             pass  # couldn't upload book. maybe blank upload?
 
     def attempt_to_update_cover(self, file):
         try:
             cover = cover_upload_set.save(file)
-            if self.cover:
-                # remove current cover if user uploaded new one
-                try:
-                    os.remove(cover_upload_set.path(self.cover))
-                except:
-                    pass  # can't delete old cover. not the end of the world.
+            if cover:
+                if self.cover:
+                    # remove current cover if user uploaded new one
+                    try:
+                        os.remove(cover_upload_set.path(self.cover))
+                    except:
+                        pass  # can't delete old cover. not the end of the world.
 
-            self.cover = cover
-            utils.create_thumbnail(cover)
+                self.cover = cover
+                utils.create_thumbnail(cover)
         except:
             pass  # couldn't upload cover. maybe blank upload?
